restructured kotlin tests
