+++ lib/uniffi-zcash/tests/kotlin/orchard/address.kts
@@ -1,0 +1,12 @@
+import uniffi.zcash.*
+
+
+fun testOrchardAddressFromRawAddressBytes() {
+	val support = TestSupport.fromCsvFile()
+	val rawBytes = support.getAsByteArray("orchard_address_bytes")
+	val zoa = ZcashOrchardAddress.fromRawAddressBytes(rawBytes)
+
+	assert(zoa.toRawAddressBytes() == rawBytes)
+
+}
+testOrchardAddressFromRawAddressBytes()
