+++ lib/uniffi-zcash/tests/kotlin/orchard/diversifier_index.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
