+++ lib/uniffi-zcash/tests/kotlin/orchard/keys/diversifier.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
