+++ lib/uniffi-zcash/tests/kotlin/orchard/keys/full_viewing_key.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
