+++ lib/uniffi-zcash/tests/kotlin/orchard/keys/scope.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
