+++ lib/uniffi-zcash/tests/kotlin/secp256k1/mod.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
