+++ lib/uniffi-zcash/tests/kotlin/zcash_client_backend/address/recipient_address.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
