+++ lib/uniffi-zcash/tests/kotlin/zcash_client_backend/address/unified_address.kts
@@ -1,0 +1,122 @@
+import uniffi.zcash.*
+
+
+fun testUnifiedAddressParsing() {
+    val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
+
+    val unifiedSpendingKey = ZcashUnifiedSpendingKey.fromSeed(
+        ZcashConsensusParameters.MAIN_NETWORK,
+        seed,
+        ZcashAccountId(0u),
+    )
+
+    val params = ZcashConsensusParameters.MAIN_NETWORK
+
+    var thrown = false;
+    try {
+        ZcashUnifiedAddress.decode(params, "")
+    } catch (e: ZcashException.Message) {
+        thrown = true;
+    }
+    assert(thrown)
+
+    val saplingDiversifier = ZcashDiversifier(listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() })
+
+    val sapling = unifiedSpendingKey.toUnifiedFullViewingKey()
+        .sapling()!!.toIvk(ZcashScope.EXTERNAL)
+        .toPaymentAddress(saplingDiversifier)
+
+    val orchardDiversifier = ZcashOrchardDiversifier.fromBytes(listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() })
+
+    val orchard = unifiedSpendingKey.toUnifiedFullViewingKey()
+        .orchard()!!.toIvk(ZcashOrchardScope.EXTERNAL)
+        .address(orchardDiversifier)
+
+    val transparent = ZcashTransparentAddress.publicKey((1..20).map { it.toUByte() })
+
+    val source = ZcashUnifiedAddress(orchard, sapling, transparent)
+    val address = source.encode(params)
+    val parsed = ZcashUnifiedAddress.decode(params, address)
+
+    assert(address == parsed.encode(params))
+}
+testUnifiedAddressParsing()
+
+fun testUnifiedAddressCreationWithSapling() {
+    val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
+
+    val unifiedSpendingKey = ZcashUnifiedSpendingKey.fromSeed(
+        ZcashConsensusParameters.MAIN_NETWORK,
+        seed,
+        ZcashAccountId(0u),
+    )
+
+    val saplingDiversifier = ZcashDiversifier(listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() })
+
+    val sapling = unifiedSpendingKey.toUnifiedFullViewingKey()
+        .sapling()!!.toIvk(ZcashScope.EXTERNAL)
+        .toPaymentAddress(saplingDiversifier)
+
+    var unifiedAddress = ZcashUnifiedAddress(null, sapling, null)
+
+    assert(sapling!!.toBytes() == unifiedAddress.sapling()!!.toBytes())
+    assert(null == unifiedAddress.orchard())
+}
+testUnifiedAddressCreationWithSapling()
+
+fun testUnifiedAddressCreation() {
+    val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
+
+    val unifiedSpendingKey = ZcashUnifiedSpendingKey.fromSeed(
+        ZcashConsensusParameters.MAIN_NETWORK,
+        seed,
+        ZcashAccountId(0u),
+    )
+
+    val saplingDiversifier = ZcashDiversifier(listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() })
+
+    val sapling = unifiedSpendingKey.toUnifiedFullViewingKey()
+        .sapling()!!.toIvk(ZcashScope.EXTERNAL)
+        .toPaymentAddress(saplingDiversifier)
+
+    val orchardDiversifier = ZcashOrchardDiversifier.fromBytes(listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() })
+
+    val orchard = unifiedSpendingKey.toUnifiedFullViewingKey()
+        .orchard()!!.toIvk(ZcashOrchardScope.EXTERNAL)
+        .address(orchardDiversifier)
+
+    val transparent = ZcashTransparentAddress.publicKey((1..20).map { it.toUByte() })
+
+    // At least one of orchard or sapling address must be set
+    // ZcashUnifiedAddress(null, null, null)
+    ZcashUnifiedAddress(orchard, null, null)
+    ZcashUnifiedAddress(null, sapling, null)
+    ZcashUnifiedAddress(orchard, sapling, null)
+    // ZcashUnifiedAddress(null, null, transparent)
+    ZcashUnifiedAddress(orchard, null, transparent)
+    ZcashUnifiedAddress(null, sapling, transparent)
+    ZcashUnifiedAddress(orchard, sapling, transparent)
+}
+testUnifiedAddressCreation()
+
+fun testUnifiedAddressCreationWithOrchard() {
+    val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
+
+    val unifiedSpendingKey = ZcashUnifiedSpendingKey.fromSeed(
+        ZcashConsensusParameters.MAIN_NETWORK,
+        seed,
+        ZcashAccountId(0u),
+    )
+
+    val orchardDiversifier = ZcashOrchardDiversifier.fromBytes(listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() })
+
+    val orchard = unifiedSpendingKey.toUnifiedFullViewingKey()
+        .orchard()!!.toIvk(ZcashOrchardScope.EXTERNAL)
+        .address(orchardDiversifier)
+
+    val unifiedAddress = ZcashUnifiedAddress(orchard, null, null)
+
+    assert(null == unifiedAddress.sapling())
+    assert(orchard.toRawAddressBytes() == unifiedAddress.orchard()!!.toRawAddressBytes())
+}
+testUnifiedAddressCreationWithOrchard()
