+++ lib/uniffi-zcash/tests/kotlin/zcash_client_backend/keys/diversifiable_full_viewing_key.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
