+++ lib/uniffi-zcash/tests/kotlin/zcash_client_backend/keys/era.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
