+++ lib/uniffi-zcash/tests/kotlin/zcash_client_backend/keys/unified_full_viewing_key.kts
@@ -1,0 +1,17 @@
+import uniffi.zcash.*
+
+fun testUnifiedFullViewingKeyEncode() {
+    val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
+
+    val unifiedSpendingKey = ZcashUnifiedSpendingKey.fromSeed(
+        ZcashConsensusParameters.MAIN_NETWORK,
+        seed,
+        ZcashAccountId(0u),
+    )
+
+    val expected = "uview1ac6swpuurz2cgr8ktk630exjrz45fsuc4jeqwgg4dm33stl8awhcju0kyaxvw58405jla4k7rqfcw35l4rsj3ta74a2me8p9hh52uxp5zm5wk60pkpy7242wdhdgm265ah3pjqe03m0vax0wa2k4yqnu0gzmnnkt2sjmxeg7s3v8j55mnrzwqznttkaj86ghs2hzp0pstlvw4zlc7kqc2n98h6xluat24829f5fvgue0w8m9r2fwtyzrdvxf7vwu67fd0wdtc0m3m952prz3w7sc8s42v48u9nsd4gld2pgjfzu9qxxxs06mdtkz2dcda0926wulk0t564k3gs6mjm04qmj6e2yrj8vmjh3flh6fg7y4k5fjj09xmv2ffv6ua7e97fszgfpp94uytsq0cu35dd53n45ua4m43gha3dquw60as4xrynllveyjczyffsd8fm6npe88pmg6j6kpjfapuurnrwjya3gz2xfvmyv5r433rsnkra8h"
+
+    assert(unifiedSpendingKey.toUnifiedFullViewingKey()
+        .encode(ZcashConsensusParameters.MAIN_NETWORK) == expected)
+}
+testUnifiedFullViewingKeyEncode()
