+++ lib/uniffi-zcash/tests/kotlin/zcash_client_backend/keys/unified_spending_key.kts
@@ -1,5 +1,21 @@
 import uniffi.zcash.*
 
+fun testUnifiedSpendingKeyFromSeed() {
+    val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
+
+    val unifiedSpendingKey = ZcashUnifiedSpendingKey.fromSeed(
+        ZcashConsensusParameters.MAIN_NETWORK,
+        seed,
+        ZcashAccountId(0u),
+    )
+
+    val expected = listOf(180, 208, 214, 194, 3, 32, 166, 3, 186, 151, 20, 139, 99, 33, 212, 134, 101, 192, 119, 208, 167, 21, 119, 228, 7, 152, 74, 140, 84, 209, 236, 235, 53, 57, 109, 65, 44, 178, 2, 169, 3, 240, 68, 53, 97, 0, 0, 0, 128, 234, 206, 224, 230, 180, 69, 172, 115, 57, 184, 221, 212, 204, 73, 161, 165, 210, 199, 46, 10, 200, 142, 73, 167, 9, 104, 89, 58, 200, 121, 136, 69, 228, 111, 114, 225, 87, 227, 210, 233, 213, 86, 13, 107, 118, 27, 114, 52, 191, 0, 154, 130, 192, 9, 11, 6, 220, 168, 246, 77, 183, 221, 52, 14, 198, 139, 75, 203, 159, 201, 17, 117, 90, 15, 68, 49, 79, 15, 95, 118, 205, 210, 120, 134, 40, 80, 122, 89, 82, 180, 159, 230, 35, 232, 105, 9, 144, 208, 234, 146, 137, 215, 60, 50, 183, 254, 149, 253, 137, 42, 232, 60, 251, 179, 135, 99, 159, 238, 119, 130, 4, 75, 67, 113, 67, 10, 191, 0, 1, 188, 15, 115, 131, 15, 198, 187, 156, 71, 47, 94, 152, 220, 110, 161, 168, 50, 123, 203, 190, 135, 6, 11, 110, 180, 235, 248, 125, 93, 89, 193, 0, 64, 207, 174, 34, 199, 252, 227, 180, 123, 230, 0, 80, 146, 93, 219, 173, 12, 108, 17, 103, 102, 144, 226, 101, 143, 138, 175, 53, 52, 12, 34, 185, 103, 172, 47, 218, 133, 127, 111, 155, 112, 212, 44, 34, 186, 124, 174, 31, 169, 99, 123, 229, 175, 141, 181, 225, 250, 107, 144, 184, 248, 64, 255, 239, 143).map { it.toUByte() }
+
+    assert(unifiedSpendingKey.toBytes(ZcashKeysEra.ORCHARD) == expected)
+}
+testUnifiedSpendingKeyFromSeed()
+
+
 fun testSpendingKeyConversions() {
     val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
 
@@ -12,7 +28,7 @@ fun testSpendingKeyConversions() {
     assert(keyBytes == expectedKeyBytes)
 
     ZcashExtendedSpendingKey.fromBytes(keyBytes)
-    
+
 }
 testSpendingKeyConversions()
 
@@ -58,7 +74,7 @@ fun testSpendingKeyDefaultAddress() {
     val key = ZcashExtendedSpendingKey.master(seed)
 
     val result = key.defaultAddress()
-    
+
     val expectedAddressBytes = listOf(91, 59, 79, 228, 216, 68, 167, 156, 199, 182, 182, 184, 107, 245, 37, 145, 194, 241, 226, 63, 157, 130, 209, 140, 137, 229, 45, 115, 56, 194, 31, 118, 140, 33, 179, 60, 74, 226, 114, 199, 101, 216, 161).map { it.toUByte() }
 
     assert( result.address.toBytes() == expectedAddressBytes)
@@ -75,7 +91,7 @@ fun testSpendingKeyDeriveInternal(){
     val key = ZcashExtendedSpendingKey.master(seed)
 
     val derivedKey = key.deriveInternal()
-    
+
     val expectedDerivedKeyBytes = listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 17, 195, 198, 202, 45, 72, 50, 2, 237, 36, 207, 134, 196, 143, 180, 241, 151, 208, 35, 90, 48, 36, 29, 242, 225, 171, 4, 154, 32, 23, 74, 135, 254, 60, 215, 251, 107, 26, 246, 149, 17, 106, 197, 95, 43, 176, 26, 114, 78, 147, 230, 102, 117, 217, 219, 121, 221, 141, 243, 105, 242, 196, 17, 6, 121, 209, 143, 180, 33, 137, 166, 142, 214, 103, 176, 171, 151, 188, 142, 87, 90, 50, 78, 79, 228, 154, 68, 183, 17, 91, 125, 90, 236, 27, 42, 6, 23, 159, 15, 53, 108, 230, 232, 135, 19, 31, 121, 130, 26, 233, 11, 243, 222, 23, 184, 46, 212, 75, 194, 186, 225, 111, 54, 239, 211, 251, 61, 87, 134, 175, 157, 85, 96, 14, 187, 153, 221, 181, 4, 120, 22, 17, 65, 203, 204, 206, 159, 46, 1, 19, 73, 177, 78, 243, 121, 114, 129, 220, 59, 205).map { it.toUByte() }
 
     assert(derivedKey.toBytes() == expectedDerivedKeyBytes)
