+++ lib/uniffi-zcash/tests/kotlin/zcash_primitives/legacy/keys/account_pub_key.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
