+++ lib/uniffi-zcash/tests/kotlin/zcash_primitives/legacy/keys/ovk.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
