+++ lib/uniffi-zcash/tests/kotlin/zcash_primitives/memo/memo_bytes.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
