+++ lib/uniffi-zcash/tests/kotlin/zcash_primitives/sapling/diversifier.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
