+++ lib/uniffi-zcash/tests/kotlin/zcash_primitives/sapling/keys/full_viewing_key.kts
@@ -1,0 +1,37 @@
+import uniffi.zcash.*
+
+fun testSaplingOvkToBytes() {
+    val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
+
+    val unifiedSpendingKey = ZcashUnifiedSpendingKey.fromSeed(
+        ZcashConsensusParameters.MAIN_NETWORK,
+        seed,
+        ZcashAccountId(0u),
+    )
+
+    val expected = listOf(144, 208, 234, 146, 137, 215, 60, 50, 183, 254, 149, 253, 137, 42, 232, 60, 251, 179, 135, 99, 159, 238, 119, 130, 4, 75, 67, 113, 67, 10, 191, 0).map { it.toUByte() }
+
+    assert(unifiedSpendingKey.toUnifiedFullViewingKey()
+        .sapling()!!.toOvk(ZcashScope.EXTERNAL)
+        .toBytes() == expected)
+}
+testSaplingOvkToBytes()
+
+fun testOrchardIvkToPaymentAddress() {
+    val seed = listOf(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() }
+
+    val unifiedSpendingKey = ZcashUnifiedSpendingKey.fromSeed(
+        ZcashConsensusParameters.MAIN_NETWORK,
+        seed,
+        ZcashAccountId(0u),
+    )
+
+    val orchardDiversifier = ZcashOrchardDiversifier.fromBytes(listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0).map { it.toUByte() })
+
+    val expected = listOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 118, 188, 133, 2, 30, 187, 222, 192, 24, 118, 136, 143, 249, 3, 127, 83, 48, 137, 67, 228, 146, 86, 27, 251, 163, 42, 159, 247, 98, 150, 25, 7).map { it.toUByte() }
+
+    assert(unifiedSpendingKey.toUnifiedFullViewingKey()
+        .orchard()!!.toIvk(ZcashOrchardScope.EXTERNAL)
+        .address(orchardDiversifier)!!.toRawAddressBytes() == expected)
+}
+testOrchardIvkToPaymentAddress()
