+++ lib/uniffi-zcash/tests/kotlin/zcash_primitives/sapling/keys/viewing_key.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
