+++ lib/uniffi-zcash/tests/kotlin/zcash_primitives/sapling/sapling_ivk.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
