+++ lib/uniffi-zcash/tests/kotlin/zcash_primitives/sapling/value/note.kts
@@ -1,0 +1,1 @@
+import uniffi.zcash.*
